Removed the ConsoleReadLine from DisplayMenu to keep methods simple and not misleading. Added a method called GetUserChoice().
Removed firstPrint variable from UI.cs and implemented it into Program.cs to maintain a constistent way of tracking the program start. This way if someone else looks into the program it will be clear.

Changed the path of QuizList.xml file to be default so it works on all systems (Windows, Mac & Linux).